Move check_broadcast from modeling.utils to utils; add tests.
